Remove webpacker
