+++ .travis.yml
@@ -1,15 +1,9 @@
 services:
   - postgresql
-before_install:
-  - curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
-  - echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
-  - sudo apt-get update
-  - sudo apt-get -y install yarn
 before_script:
   - cp config/database.yml.travis config/database.yml
   - psql -c 'create database travis_ci_test;' -U postgres
   - bundle exec rails db:migrate
-  - nvm install 12.18.4
   - bundle exec rake assets:precompile
 language: ruby
 rvm:
