--- app/javascript/app/index.js
@@ -1,3 +1,0 @@
-import { App } from './app';
-
-App.initApp();
