--- app/javascript/app/styles/app.scss
@@ -1,4 +1,0 @@
-@import '~materialize-css/sass/components/color-variables';
-$primary-color: color('blue', 'darken-1') !default;
-$secondary-color: color('deep-orange', 'base') !default;
-@import '~materialize-css/sass/materialize';
