--- app/javascript/packs/stylesheets.scss
@@ -1,1 +1,0 @@
-@import '../app/styles/app';
