+++ app/views/layouts/application.html.erb
@@ -7,8 +7,6 @@
 
     <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= javascript_include_tag 'application' %>
-    <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
-    <%= stylesheet_pack_tag 'stylesheets' %>
     <meta name="viewport" content="width=device-width, initial-scale=1">
   </head>
