--- config/webpack/environment.js
@@ -1,3 +1,0 @@
-const { environment } = require('@rails/webpacker')
-
-module.exports = environment
