--- package.json
@@ -1,19 +1,0 @@
-{
-  "name": "flight_cards",
-  "private": true,
-  "dependencies": {
-    "@rails/actioncable": "^6.0.0",
-    "@rails/activestorage": "^6.0.0",
-    "@rails/ujs": "^6.0.0",
-    "@rails/webpacker": "4.3.0",
-    "materialize-css": "^1.0.0-rc.2",
-    "turbolinks": "^5.2.0"
-  },
-  "version": "0.1.0",
-  "devDependencies": {
-    "webpack-dev-server": "^3.11.0"
-  },
-  "resolutions": {
-    "node-forge": "0.10.0"
-  }
-}
